+++ CHANGELOG.md
@@ -31,8 +31,16 @@ Thank you to all who have contributed!
 ### Added
 - Adds `org.partiql.value` (experimental) package for reading/writing PartiQL
   values
+- Adds serialization and deserialization between IonValue and `org.partiql.value`.
+- Adds `org.partiql.ast` package and usage documentation
+- Adds `org.partiql.parser` package and usage documentation
+- Adds PartiQL's Timestamp Data Model. 
+- Adds support for Timestamp constructor call in Parser. 
+
 
 ### Changed
+
+- Upgrades ion-java to 1.10.2.
 
 ### Deprecated
 
@@ -45,6 +53,7 @@ Thank you to all who have contributed!
 ### Contributors
 Thank you to all who have contributed!
 - @howero
+- @yliuuuu
 - @<your-username>
 
 ## [0.12.0] - 2023-06-14